mailserver/requirements.txt
salmon-mail
celery
